создал поддиректории под каждый view
