+++ src/view/point-view/point-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js'
+import { createElement } from '../../render.js'
 
 function createPointTemplate() {
   return `
