+++ src/view/sort-view/sort-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js'
+import { createElement } from '../../render.js'
 
 function createSortTemplate() {
   return ` <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -29,7 +29,7 @@ function createSortTemplate() {
           </form>`;
 }
 
-export default class Sort {
+export default class SortView {
   getTemplate() {
     return createSortTemplate()
   }
